pyproject.toml
[project]
name = "drtail-receipt-scanner-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "black>=23.0.0",
  "google-generativeai>=0.3.0",
  "mistralai>=1.0.0",
  "mypy>=1.7.0",
  "openai>=1.12.0",
  "pandas>=2.0.0",
  "pillow>=10.0.0",
  "pydantic>=2.5.0",
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "python-dotenv>=1.0.0",
  "pyyaml>=6.0",
  "requests>=2.31.0",
  "ruff>=0.1.0",
  "streamlit>=1.30.0",
  "veryfi>=3.0.0",
]
